+++ app/Api/Resources/RefundResource.php
@@ -2,13 +2,12 @@
 
 namespace App\Api\Resources;
 
-use App\Api\Repositories\RefundRepository;
 use Illuminate\Contracts\Support\Arrayable;
 use Illuminate\Http\Request;
 use Illuminate\Http\Resources\Json\JsonResource;
 use JsonSerializable;
 
-class CheckoutResource extends JsonResource
+class RefundResource extends JsonResource
 {
     /**
      * Transform the resource into an array.
@@ -18,14 +17,16 @@ class CheckoutResource extends JsonResource
      */
     public function toArray($request)
     {
+        $response = json_decode($this->response);
+
         return [
-            'id' => $this->checkout_id,
+            'id' => $this->refund_id,
             'uuid' => $this->uuid,
-            'amount' => $this->amount,
-            'reference' => $this->reference,
-            'status' => $this->status,
-            'payment' => new PaymentResource($this->payment),
-            'refunds' => RefundResource::collection($this->payment->refund),
+            'amount' => $response->amount,
+            'currency' => $response->currency,
+            'code' => $response->result->code,
+            'description' => $response->result->description,
+            'completedAt' => $this->completed_at,
         ];
     }
 }