+++ CHANGELOG.md
@@ -10,6 +10,6 @@
 - ADDED migrations and models for accounts
 - ADDED requests create user
 - ADDED resources user, account types
-- ADDED create user, get user, auth endpoint, create checkout, get checkouts
+- ADDED create user, get user, auth endpoint, create checkout, get checkouts, create refund
 - ADDED Bearer Tokens
 - ADDED Oppwa integration